refactor ast functions out of backend compiler.
