+++ ivy/utils/__init__.py
@@ -1,2 +1,4 @@
 # flake8: noqa
 from . import backend
+from . import dynamic_import
+from .dynamic_import import *
