+++ ivy/utils/backend/__init__.py
@@ -2,3 +2,4 @@
 from . import handler
 from .handler import *
 from . import compiler
+from ._importlib import _ivy_import_module
