+++ ivy/utils/dynamic_import.py
@@ -1,0 +1,4 @@
+# NOQA
+from importlib import import_module as imp_md
+
+import_module = imp_md
